+++ examples/serve-indexes.ts
@@ -3,20 +3,15 @@ import { join } from 'path';
 
 import express from 'express';
 
-import { FileSystemStorage, TrailingSlashError } from '../src/send-stream';
+import { FileSystemStorage } from '../src/send-stream';
 
 const app = express();
 
-const storage = new FileSystemStorage(join(__dirname, 'assets'));
+const storage = new FileSystemStorage(join(__dirname, 'assets'), { onDirectory: 'serve-index' });
 
 app.get('*', async (req, res, next) => {
 	try {
-		let result = await storage.prepareResponse(req.url, req);
-		// if the path is not found and the reason is a trailing slash then try to load matching index.html
-		if (result.error instanceof TrailingSlashError) {
-			result.stream.destroy();
-			result = await storage.prepareResponse([...result.error.pathParts.slice(0, -1), 'index.html'], req);
-		}
+		const result = await storage.prepareResponse(req.url, req);
 		result.send(res);
 	} catch (err) {
 		// eslint-disable-next-line node/callback-return